Well, TESR works, also, Cleaned up a litle

+++ src/main/java/EmeraldCasino/renderers/blocks/TileEntityCardBlockRenderer.java
@@ -2,12 +2,10 @@ package EmeraldCasino.renderers.blocks;
 
 import org.lwjgl.opengl.GL11;
 
-import cpw.mods.fml.client.registry.ISimpleBlockRenderingHandler;
 import cpw.mods.fml.client.registry.RenderingRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.OpenGlHelper;
-import net.minecraft.client.renderer.RenderBlocks;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.client.renderer.tileentity.TileEntitySpecialRenderer;
 import net.minecraft.entity.Entity;
@@ -62,7 +60,7 @@ public class TileEntityCardBlockRenderer extends TileEntitySpecialRenderer {
          int modulousModifier = skyLight % 65536;
          int divModifier = skyLight / 65536;
          tess.setColorOpaque_F(brightness, brightness, brightness);
-         OpenGlHelper.setLightmapTextureCoords(OpenGlHelper.lightmapTexUnit,  (float) modulousModifier,  divModifier);
+         OpenGlHelper.setLightmapTextureCoords(OpenGlHelper.lightmapTexUnit,  modulousModifier,  divModifier);
  }
 
 }
